UserController: sends a response if user has not been created prevent auth user to delete, restore or destory herself
UserResource:
add relation avatar to UserResource

User:
remove comments method for it and morphed the avatar method

	modified:   app/Http/Controllers/Api/UserController.php
	modified:   app/Http/Resources/UserResource.php
	modified:   app/Models/User.php
	modified:   app/Repositories/UserRepository.php

+++ app/Http/Controllers/Api/UserController.php
@@ -12,6 +12,7 @@ use App\Repositories\Facades\UserRepo;
 use App\Http\Resources\PermissionCollection;
 use App\Http\Requests\User\StoreUserRequest;
 use App\Http\Requests\User\UpdateUserRequest;
+use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
@@ -22,16 +23,22 @@ class UserController extends Controller
 
     public function store(StoreUserRequest $request)
     {
-        $createdUser   =   UserRepo::create($request->except('avatar'));
+        define('NOT_CREATED', 0);
+        $createdUser    =   UserRepo::create($request->except('avatar'));
+        if ($createdUser === NOT_CREATED) {
+            return new DBResource($createdUser);
+        }
+
         if ($request->hasFile('avatar')) {
             $createdUser->addMediaFromRequest('avatar')->toMediaCollection(enum('media.user.avatar'));
         }
+
         return new UserResource($createdUser);
     }
 
     public function show(User $user)
     {
-        return new UserResource($user);
+        return new UserResource($user->load('avatar'));
     }
 
     public function update(UpdateUserRequest $request, User $user)
@@ -44,16 +51,55 @@ class UserController extends Controller
 
     public function delete(User $user)
     {
+        if (auth_user()->is($user)) {
+            return response(
+                [
+                    'error' => [
+                        'unauthorized'  =>  trans('http.unauthorized')
+                    ]
+                ],
+                Response::HTTP_UNAUTHORIZED,
+                [
+                    'Content-Type' => enum('system.response.json')
+                ]
+            );
+        }
         return new DBResource(UserRepo::delete($user));
     }
 
     public function restore(string $user)
     {
+        if (auth_user()->id == $user) {
+            return response(
+                [
+                    'error' => [
+                        'unauthorized'  =>  trans('http.unauthorized')
+                    ]
+                ],
+                Response::HTTP_UNAUTHORIZED,
+                [
+                    'Content-Type' => enum('system.response.json')
+                ]
+            );
+        }
         return new DBResource(UserRepo::restore($user));
     }
 
     public function destroy(string $user)
     {
+        if (auth_user()->id == $user) {
+            return response(
+                [
+                    'error' => [
+                        'unauthorized'  =>  trans('http.unauthorized')
+                    ]
+                ],
+                Response::HTTP_UNAUTHORIZED,
+                [
+                    'Content-Type' => enum('system.response.json')
+                ]
+            );
+        }
         return new DBResource(UserRepo::destroy($user));
     }
